+++ src/main.c
@@ -7,8 +7,15 @@
 int
 main(int argc, char **argv)
 {
-	VipsdispApp *app;
-	int status;
+#ifdef FLATPAK
+    /* In flatpak builds, don't pick up VIPSHOME from the environ, we want the
+     * value detected for this install.
+     */
+    g_unsetenv("VIPSHOME");
+#endif /*FLATPAK*/
+
+	// disable DoS limits on libvips 8.19+
+    g_setenv("VIPS_UNLIMITED", "1", TRUE);
 
 	if (VIPS_INIT(argv[0]))
 		vips_error_exit("unable to start libvips");
@@ -28,9 +35,9 @@ main(int argc, char **argv)
 	g_setenv("G_DEBUG", "fatal-warnings", FALSE);
 #endif /*DEBUG*/
 
-	app = vipsdisp_app_new();
+	VipsdispApp *app = vipsdisp_app_new();
 
-	status = g_application_run(G_APPLICATION(app), argc, argv);
+	int status = g_application_run(G_APPLICATION(app), argc, argv);
 
 	vips_shutdown();